requirements.txt
tiktoken~=0.4.0
nltk~=3.8.1
cassandra-driver~=3.28.0
python-dateutil~=2.8.2
scikit-learn~=1.3.0
numpy~=1.25.0
google-generativeai~=0.7.2
python-fasthtml~=0.2.4
xxhash~=3.4.1
requests~=2.32.3
